refactor: move runtime capabilities into core
