+++ crates/shelterwood-core/src/proxied_sleep.rs
@@ -1,0 +1,135 @@
+use std::{
+    future::Future,
+    pin::Pin,
+    sync::Arc,
+    task::{Context, Poll},
+};
+
+use crate::{
+    BoxedSleep, MailboxRuntime,
+    panic::PanicAccumulator,
+    waker::{WakerAction, WakerEffects},
+    waker_proxy::ProxiedPoll,
+};
+
+/// Timer future that keeps a caller-owned waker out of the runtime's wheel.
+///
+/// Runtime adapters supply the raw timer, while this wrapper owns the common
+/// framework boundary: the external primitive registers only a stable proxy,
+/// and the caller's waker stays in the proxy's effects-mediated private slot.
+/// A timer that is ready on the first no-op probe allocates no proxy.
+///
+/// Poll-path retirement is synchronous and contained. Drop-glue retirement
+/// first hands the caller waker to the runtime's disposal lane, then cancels
+/// the framework-only wheel entry synchronously. This is the venue split
+/// established for public mailbox deadlines by #398.
+#[doc(hidden)]
+pub struct ProxiedSleep {
+    runtime: Arc<dyn MailboxRuntime>,
+    timer: Option<BoxedSleep>,
+    timer_poll: ProxiedPoll,
+    completed: bool,
+}
+
+impl ProxiedSleep {
+    /// Wraps one raw runtime timer.
+    ///
+    /// The timer must be framework-owned: its poll and drop implementations
+    /// are invoked inside framework containment boundaries, and the supported
+    /// façade exposes no way for an application to construct this type.
+    #[doc(hidden)]
+    pub fn new(timer: BoxedSleep, runtime: Arc<dyn MailboxRuntime>) -> Self {
+        Self {
+            runtime,
+            timer: Some(timer),
+            timer_poll: ProxiedPoll::new(),
+            completed: false,
+        }
+    }
+
+    /// Retires an armed timer on the polling thread.
+    ///
+    /// Used when a containing future completes by a non-timer branch. The
+    /// caller-waker destructor is fully drained before the containing future
+    /// hands its result back, and the wheel entry is synchronously removed.
+    #[doc(hidden)]
+    pub fn retire_inline(&mut self, panics: &mut PanicAccumulator) {
+        self.retire(WakerAction::DropInline, panics);
+    }
+
+    /// Cancels this timer on a containing future's successful poll path.
+    ///
+    /// Public only as a sibling-crate implementation seam. Runtime selection
+    /// helpers call it when a non-timer branch wins, so that semantic poll
+    /// path retains the synchronous-contained half of the venue split even
+    /// though the timer itself is the losing future. Retirement fuses the
+    /// sleep: a later poll reports ready rather than reaching for the
+    /// emptied timer slot.
+    #[doc(hidden)]
+    pub fn cancel_inline(&mut self) {
+        let mut panics = PanicAccumulator::default();
+        self.retire_inline(&mut panics);
+        crate::panic::discard_panic(panics.take());
+    }
+
+    fn retire_disposing(&mut self, panics: &mut PanicAccumulator) {
+        self.retire(WakerAction::Dispose(Arc::clone(&self.runtime)), panics);
+    }
+
+    fn retire(&mut self, action: WakerAction, panics: &mut PanicAccumulator) {
+        let mut effects = WakerEffects::default();
+        let timer_poll = &mut self.timer_poll;
+        // The proxy mutex guards framework-owned data and is documented
+        // unpoisonable, but this path also runs during unwind. Keep every
+        // cleanup step inside the accumulator so a bookkeeping defect cannot
+        // turn a caller panic into an abort.
+        panics.run(|| timer_poll.retire(action, &mut effects));
+        effects.flush(panics);
+
+        // Slot first, timer second: once the caller waker is gone, cancelling
+        // the wheel entry can deliver no stale wake to a caller that already
+        // has its answer. The timer still drops synchronously so the entry is
+        // gone when retirement returns.
+        let timer = self.timer.take();
+        panics.run(|| drop(timer));
+        // Retirement is the single point that empties both slots, so it also
+        // fuses the future: without this, a poll after `cancel_inline` would
+        // panic on the emptied timer slot instead of reporting ready.
+        self.completed = true;
+    }
+}
+
+impl Future for ProxiedSleep {
+    type Output = ();
+
+    fn poll(mut self: Pin<&mut Self>, context: &mut Context<'_>) -> Poll<()> {
+        let this = self.as_mut().get_mut();
+        if this.completed {
+            return Poll::Ready(());
+        }
+        // `ProxiedPoll::poll` retires the caller slot inline on the ready
+        // edge; `retire_inline` below then finds it empty and only cancels
+        // the wheel entry, preserving the slot-first, timer-second order.
+        let result = this.timer_poll.poll(
+            this.timer
+                .as_mut()
+                .expect("an incomplete proxied sleep retains its timer"),
+            context,
+            |timer, context| timer.as_mut().poll(context),
+            Poll::is_pending,
+        );
+        if result.is_ready() {
+            let mut panics = PanicAccumulator::default();
+            this.retire_inline(&mut panics);
+            crate::panic::discard_panic(panics.take());
+        }
+        result
+    }
+}
+
+impl Drop for ProxiedSleep {
+    fn drop(&mut self) {
+        let mut panics = PanicAccumulator::default();
+        self.retire_disposing(&mut panics);
+    }
+}
