+++ crates/shelterwood-core/src/waker_proxy.rs
@@ -5,8 +5,8 @@ use std::{
 };
 
 use crate::{
-    cell::waker_slot::{WakerAction, WakerEffects, WakerSlot},
     panic::PanicAccumulator,
+    waker::{WakerAction, WakerEffects, WakerSlot},
 };
 
 /// Reusable probe, registration, and proxy-poll state machine.
@@ -91,14 +91,9 @@ impl ProxiedPoll {
         crate::panic::discard_panic(panics.take());
     }
 
-    /// Test-only visibility: whether a pending poll has a proxy installed.
-    #[cfg(test)]
-    pub(crate) fn is_parked(&self) -> bool {
-        self.proxy.is_some()
-    }
-
     /// Retires the current caller registration into a mailbox effects sink.
-    pub(crate) fn retire(&mut self, action: WakerAction, effects: &mut WakerEffects) {
+    #[doc(hidden)]
+    pub fn retire(&mut self, action: WakerAction, effects: &mut WakerEffects) {
         let proxy = self.proxy.take();
         if let Some(proxy) = &proxy {
             proxy.retire(action, effects);
@@ -309,7 +304,7 @@ mod tests {
     };
 
     use super::{WakerProxy, WakerProxyState};
-    use crate::cell::waker_slot::{WakerAction, WakerEffects};
+    use crate::waker::{WakerAction, WakerEffects};
 
     #[derive(Default)]
     struct CountWake(AtomicUsize);
